+++ src/lib/common/XMPPUtils.ts
@@ -39,6 +39,20 @@ export class XMPPUTils {
         return messageToSendID;
     }
 
+    getUniqueId(suffix) {
+        let uuid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
+            let r = Math.random() * 16 | 0,
+                v = c === 'x' ? r : r & 0x3 | 0x8;
+            return v.toString(16);
+        });
+
+        if (typeof suffix === "string" || typeof suffix === "number") {
+            return uuid + ":" + suffix;
+        } else {
+            return uuid + "";
+        }
+    }
+
     generateRandomFullJidForNode(jid, generatedRandomId) {
         let fullJid = jid + "/node_" + generatedRandomId;
         return fullJid;